menampilkan kerja kami pada public

+++ admin/publikasi-delete.php
@@ -1,7 +1,14 @@
 <?php 
 require 'koneksi.php';
 
-$id = $_POST["hapus_data"];
+session_start();
+// cek apakah sudah login
+if (!isset($_SESSION["login"])) {
+    header("Location: login.php");
+    exit;
+}
+
+$id = $_GET["hapus_data"];
 
 $hapus_data = "DELETE FROM publikasi WHERE id = '$id'";
 $hapus_berhasil = mysqli_query($conn, $hapus_data);
@@ -12,10 +19,12 @@ if($hapus_berhasil){
     alert('Data berhasil dihapus');
     document.location='publikasi.php';
     </script>";
+    exit;
 } else {
     echo"<script> 
     alert('Data gagal dihapus');
     document.location='publikasi.php';
     </script>";
+    exit;
 }
 ?>
